requirements.txt
tensorflow==2.14.0
Image
Pillow
cairosvg
numpy
tensorflow-macos
tensorflow-metal
